### 2023-07-10 #5638 @xuanxu
np, I want to improve the bot to be more resilient to this kind of "errors"
