### 2023-07-10 f002571
Apply suggestions to README
Thanks to suggestions by @Kevin-Mattheus-Moerman.
