### 2023-07-10 #5638 @rikhuijzer
> @rikhuijzer thanks for this submission. I am one of the AEiCs for JOSS and here to help process initial steps. I have started to look for a handling editor. I quickly had a look at your repository. I recommend you extend your README with the usual information like a (somewhat lay-person) summary of what the software is about, some quick information/links on how to install/get started, and then some contributing/community guidelines as well.

Thank you for the suggestions, @Kevin-Mattheus-Moerman. I've updated the [`README`](https://github.com/rikhuijzer/SIRUS.jl).
