### 2023-07-18 1bf1892
Remove deprecated `docstring` keyword (#38)
@rikhuijzer The docstring keyword is deprecated. see
[here](https://github.com/JuliaAI/MLJModelInterface.jl/blob/dev/src/metadata_utils.jl#L130)
