### 2023-07-18 64b538b
Add `Operations` and `Fitted parameters` sections (#39)
Based on feedback from https://github.com/JuliaAI/MLJModels.jl/pull/519.
