### 2023-07-18 e18a64c
Make regressors `Deterministic` (fix #40)
