+++ src/SIRUS.jl
@@ -9,7 +9,7 @@ using InlineStrings: String255
 using LinearAlgebra: rank
 using MLJBase: mode
 using MLJLinearModels: MLJLinearModels, RidgeRegressor, glr
-using MLJModelInterface: UnivariateFinite, Probabilistic, fit
+using MLJModelInterface: Deterministic, UnivariateFinite, Probabilistic, fit
 using Random: AbstractRNG, default_rng, seed!, shuffle
 using Statistics: mean, median
 using Tables: Tables, matrix
