### 2023-08-14 #5638 @rikhuijzer
In a discussion with a colleague, it turned out that one of the main arguments in favor of the SIRUS algorithm could be improved. Since the paper was still on the waitlist, I've updated that in [15b29fa](https://github.com/rikhuijzer/SIRUS.jl/commit/15b29fa512c51dbb68459e77a65f4c87763a072b). Let's generate a new PDF:
