### 2023-08-17 #5638 @jbytecode
Dear @rikhuijzer, I am the handling editor of this submission. Thank you for submitting your work to JOSS.

Do you have suggestions for suitable reviewers? You can use the link (https://reviewers.joss.theoj.org/lookup) for searching and filtering suitable reviewers. Please mention their GitHub usernames without using the `@` character just because we don't want to bother them. If you don't have a set of suggestions, I will try to find reviewers for you.

Thank you in advance.
