### 2023-08-18 #5638 @rikhuijzer
Dear @jbytecode, thank you for looking into our paper and accepting to be the editor.

> Do you have suggestions for suitable reviewers?

Based on going through the list, here are some suggestions:

`sylvaticus` could be a suitable reviewer. He has experience with Julia, machine learning for smaller datasets (<10k datapoints), and has also implemented the MLJ interface for his BetaML.jl package.

Next, `juliohm` could be a suitable reviewer since he has authored many Julia packages satisfying the MLJ interface.

Furthermore, `gdalle` could be a suitable reviewer. I've seen and use some of his projects. Next to Julia experience, he also has machine learning experience.

Finally, `pitmonticone` could also be a suitable reviewer. He has a lot of experience in statistical models and modelling in Julia including some COVID data modelling, NLP, and some time series projects.
