### 2023-08-18 #5638 @jbytecode
👋👋👋 Dear @sylvaticus and Dear @juliohm 👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**SIRUS.jl: Interpretable Machine Learning via Rule Extraction**". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/5638).

Thank you in advance!
