### 2023-08-18 #5638 @juliohm
Dear all, unfortunately too busy at the moment to review. Thanks for
checking in.

Em sex., 18 de ago. de 2023 14:03, Mehmet Hakan Satman <
***@***.***> escreveu:

> 👋👋👋 Dear @sylvaticus <https://github.com/sylvaticus> and Dear @juliohm
> <https://github.com/juliohm> 👋👋👋
>
> Would you be willing to assist in reviewing this submission for *JOSS* (*Journal
> of Open Source Software*)?
>
> JOSS publishes articles about open source research software. The
> submission I'd like you to review is titled: "*SIRUS.jl: Interpretable
> Machine Learning via Rule Extraction*". You can find more information at
> the top of this Github issue (#5638
> <https://github.com/openjournals/joss-reviews/issues/5638>).
>
> The review process at JOSS is unique: it takes place in a GitHub issue, is
> open, and author-reviewer-editor conversations are encouraged. If you have
> any questions please let me know.
>
> This is the pre-review issue. After setting at least 2 reviewers we will
> start the review process in a separate thread. In that thread, there will
> be 23 check items for each single reviewer.
>
> Thank you in advance!
>
