### 2023-08-22 #5638 @sylvaticus
Hello, here are my first comments on a quick "first passage" of the paper itself (I haven't yet installed the software). Most are minor points.

- perhaps add two words on the concept of interpretable model in the Statement of needs section?
- "black box models may lead to unsafe or unreliable prediction". I would add also "unfair" :-)
- in the notes to the table with the fitted model I would use 1/0 as classes (instead of 1.0 and 0.0) and specify that class 1 is survival, e.g. " showing only the probability for class 1 (survival) since class 0 has probability 1-p"
- in Table 2 I would add a column with the kind of prediction of each model: binary classification/multiclass classification/regression (perhaps to save space with abbreviations like BC/MC/R ??)
- not clear in the paper the suggestion to use SymbolicRegression.jl
- as the core is on interpretability, you could show in the code example also how to obtain the fitted rules

Aside of that, what is not clear in the paper is if SUIRUS.jl depends from a Regression Forest package, if it implements a RF algorithm itself (and in such cases how we move from RF to the "simple rules") or if the algorithm is independent on it, and in such cases which are the key procedures to determine/fit the "simple rules" (it doesn't need to be exhaustive, that's not the scope of a JOSS paper I believe, just so that the reader has an overview).
