### 2023-08-24 383b25f
Switch docs to EvoTrees.jl
