### 2023-08-24 27d932e
Switch docs to EvoTrees.jl (#45)
LightGBM still doesn't work (see
https://github.com/rikhuijzer/SIRUS.jl/pull/34 for more info) and
XGBoost crashed Pluto. EvoTrees did work.
