+++ docs/Project.toml
@@ -6,7 +6,7 @@ DataDeps = "124859b0-ceae-595e-8997-d05f6a7a8dfe"
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 DecisionTree = "7806a523-6efd-50cb-b5f6-3fa6f1930dbb"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
-LightGBM = "7acf609c-83a4-11e9-1ffb-b912bcd3b04a"
+EvoTrees = "f6006082-12f8-11e9-0c9c-0d5d367ab1e5"
 LiveServer = "16fef848-5104-11e9-1b77-fb7a48bbb589"
 MLJ = "add582a8-e3ab-11e8-2d5e-e98b27df1bc7"
 MLJBase = "a7f614a8-145f-11e9-1d2a-a57a1082229d"
@@ -25,6 +25,7 @@ DataDeps = "0.7"
 DataFrames = "1"
 DecisionTree = "0.12"
 Documenter = "0.27"
+EvoTrees = "0.16"
 LiveServer = "1"
 MLJ = "0.19"
 MLJBase = "0.21"
