### 2023-08-24 9c65124
Allow `Int` classes
