+++ paper/paper.bib
@@ -9,7 +9,7 @@
   doi={10.1016/j.inffus.2019.12.012}
 }
 
-@article{benard2021interpretable,
+@article{benard2021sirus,
 title={{SIRUS: Stable and Interpretable RUle Set for classification}},
 author={Cl{\'e}ment B{\'e}nard and G{\'e}rard Biau and S{\'e}bastien Da Veiga and Erwan Scornet},
 volume={15},
@@ -20,6 +20,15 @@ pages={427 -- 505},
 year={2021},
 doi={10.1214/20-EJS1792},
 URL={https://doi.org/10.1214/20-EJS1792}
+}
+
+@inproceedings{benard2021interpretable,
+  title={Interpretable random forests via rule extraction},
+  author={B{\'e}nard, Cl{\'e}ment and Biau, G{\'e}rard and Da Veiga, S{\'e}bastien and Scornet, Erwan},
+  booktitle={International Conference on Artificial Intelligence and Statistics},
+  pages={937--945},
+  year={2021},
+  organization={PMLR}
 }
 
 @article{biau2016random,
