+++ paper/build.sh
@@ -6,4 +6,5 @@ docker run --rm \
     --volume $DIR:/data \
     --user $(id -u):$(id -g) \
     --env JOURNAL=joss \
+    --platform linux/amd64 \
     openjournals/inara
