### 2023-08-25 0b9d13b
Make check more robust
