### 2023-08-25 89e3026
Allow `Int` classes (#46)
Antonello Lobianco
[noticed](https://github.com/openjournals/joss-reviews/issues/5638#issuecomment-1688233929)
that the outcome classes are typically floats (for example, `1.0` and
`0.0`) whereas integers would be much more suitable. For example, this
was the output for the haberman dataset:
```
StableRules model with 8 rules:
 if X[i, :nodes] < 8.0 then 0.156 else 0.031 +
 if X[i, :nodes] < 14.0 then 0.164 else 0.026 +
 if X[i, :nodes] < 4.0 then 0.128 else 0.037 +
 if X[i, :nodes] ≥ 8.0 & X[i, :age] < 38.0 then 0.0 else 0.008 +
 if X[i, :year] ≥ 1966.0 & X[i, :age] 
[…]
