+++ paper/paper.bib
@@ -180,6 +180,14 @@ URL={https://doi.org/10.1214/20-EJS1792}
   organization={American Medical Informatics Association}
 }
 
+@article{taleb2020statistical,
+  title={Statistical consequences of fat tails: Real world preasymptotics, epistemology, and applications},
+  author={Taleb, Nassim Nicholas},
+  journal={arXiv preprint arXiv:2001.10488},
+  year={2020},
+  doi={10.48550/arXiv.2001.10488}
+}
+
 @misc{wolberg1995breast,
   author={Wolberg, William, Mangasarian, Olvi, Street, Nick, and Street, W.},
   title={{Breast Cancer Wisconsin (Diagnostic)}},
