+++ paper/paper.md
@@ -58,9 +58,9 @@ We have also set the maximum depth hyperparameter to 2.
 This hyperparameter means that the random forests inside the algorithm are not allowed to have a depth higher than 2.
 In turn, this means that rules contain at most 2 clauses (`if A & B`).
 When the maximum depth is set to 1, then the rules contain at most 1 clause (`if A`).
-Most rule-based models, including SIRUS, are restricted to depth of 1 or 2 [@beneard2021sirus].
+Most rule-based models, including SIRUS, are restricted to depth of 1 or 2 [@benard2021sirus].
 
-The fitted model looks as follows (see Section _Code Example_ for the code):
+The output for the fitted model looks as follows (see Section _Code Example_ for the code):
 
 ```
 StableRules model with 8 rules:
@@ -77,7 +77,7 @@ Note: showing only the probability for class 1 since class 0 has
       probability 1 - p.
 ```
 
-The output shows that the model contains 8 rules.
+This shows that the model contains 8 rules.
 The first rule, for example, can be interpreted as:
 
 _If the number of detected axillary nodes is lower than 8, then take 0.156, otherwise take 0.031._
@@ -85,19 +85,19 @@ _If the number of detected axillary nodes is lower than 8, then take 0.156, othe
 This calculation is done for all 8 rules and the score is summed to get a prediction.
 In essence, the first rule says that if there are less than 8 axillary nodes detected, then the patient will most likely survive (`class == 1.0`).
 Put differently, the model states that if there are many axillary nodes detected, then it is, unfortunately, less likely that the patient will survive.
-This model is fully interpretable because there are few rules which can all be interpreted in isolation and together.
-In contrast, Random forests consist of hundreds to thousands of trees, which makes them too complex to interpret.
+This model is fully interpretable because the model contains a few dozen rules which can all be interpreted in isolation and together.
+In contrast, Random forests require hundreds to thousands of trees, which makes them too complex to interpret.
 
 # Stability
 
-Another problem that the SIRUS algorithm improves upon is that of model stability.
+Another problem that the SIRUS algorithm addresses is that of model stability.
 A stable model is defined as a model which leads to similar conclusions for small changes to data [@yu2020veridical].
 Unstable models can be difficult to apply in practice as they might require processes to constantly change.
 This also makes such models appear less trustworthy.
 Put differently, an unstable model by definition leads to different conclusions for small changes to the data and, hence, small changes to the data could cause a sudden drop in predictive performance.
 One model which suffers from a low stability is a decision tree because it will first create the root node of the tree, so a small change in the data can cause the root, and therefore the rest, of the tree to be completely different [@molnar2022interpretable].
 Similarly, linear models can be highly sensitive to correlated data and, in the case of regularized linear models, the choice of hyperparameters.
-Instead, the SIRUS algorithm provides stability by stabilizing the trees and the authors have proven the correctness of this stabilization mathematically [@beneard2021sirus].
+Instead, the SIRUS algorithm provides stability by stabilizing the trees and the authors have proven the correctness of this stabilization mathematically [@benard2021sirus].
[…]
