### 2023-08-25 6f39b12
Remove non-paper workflows
