+++ .github/workflows/CI.yml
@@ -1,36 +1,0 @@
-name: CI
-
-on:
-  push:
-    branches:
-      - main
-    paths-ignore:
-      - "README.md"
-      - "LICENSE.md"
-  pull_request:
-  workflow_dispatch:
-
-jobs:
-  Test:
-    name: Test
-    runs-on: ${{ matrix.os }}
-    timeout-minutes: 30
-    strategy:
-      fail-fast: false
-      matrix:
-        version:
-          - '1.6'
-          - '1'
-        os:
-          - ubuntu-latest
-
-    steps:
-      - uses: actions/checkout@v3
-      - uses: julia-actions/setup-julia@v1
-        with:
-          version: ${{ matrix.version }}
-      - uses: julia-actions/cache@v1
-      - uses: julia-actions/julia-buildpkg@v1
-      - uses: julia-actions/julia-runtest@v1
-        with:
-          coverage: false
