### 2023-08-26 30e499f
Fix two mistakes in the text
