### 2023-08-26 #5638 @rikhuijzer
Thank you very much @sylvaticus for your comments! The comment about the classes not being shown as `1` and `0`, but instead as `1.0` and `0.0` was especially a good one and it is fixed that now in the SIRUS code (https://github.com/rikhuijzer/SIRUS.jl/pull/46). We've also incorporated your other comments in the paper and are happy with the improved clarity of the text.
