### 2023-08-29 #5786 @editorialbot
[REVIEW]: SIRUS.jl: Interpretable Machine Learning via Rule Extraction
**Submitting author:** @rikhuijzer (<a href="http://orcid.org/0000-0001-9445-8466">Rik Huijzer</a>)
**Repository:** https://github.com/rikhuijzer/SIRUS.jl
**Branch with paper.md** (empty if default branch): rh/paper
**Version:** v1.3.3
