### 2023-09-07 rikhuijzer/SIRUS.jl#47 @gdalle
Remarks on the binary classification tutorial
As part of my JOSS review, here is a first batch of comments on the tutorial:

> auxillary nodes (Random forests & other sections)

axillary nodes, presumably

> The model has learned three rules for this dataset. (Interpretation)

Isn't it 8?

> This is done for all the rules and, finally, the rules are summed to obtain the final prediction. (Interpretation)

It took me a while to understand that they are weighted first. At the very least, saying average instead of sum would be clearer, but weighted average would be best.

By the way, do the weights sum to 1? And how are they obtained, since frequency ranking of the rules is already leveraged for pruning the ruleset?

> The x-position on the left shows `log(else-scores / if-scores)` (Visualization)

Not sure what is meant here. I don't really understand the plots as a result.

> all rules (fitted in the different cross-validation folds)

Shouldn't there be $8f$ rules, where $f$ is the number of folds? I see much fewer

https://github.com/openjournals/joss-reviews/issues/5786
