### 2023-09-07 rikhuijzer/SIRUS.jl#48 @gdalle
Inconsistent numerical results
The numerical results reported in the JOSS paper are inconsistend with the latest CI benchmark https://github.com/rikhuijzer/SIRUS.jl/actions/runs/5972815140

https://github.com/openjournals/joss-reviews/issues/5786
