### 2023-09-07 rikhuijzer/SIRUS.jl#49 @gdalle
Simple example with prediction
The tutorial is extremely detailed, but it benchmarks a handful of approaches before even showing how to use the rule-based classifier. Maybe a dead simple example of fit-predict-evaluate would belong on the front page of the docs + in the paper, for those unfamiliar with the MLJ framework

https://github.com/openjournals/joss-reviews/issues/5786
