### 2023-09-07 rikhuijzer/SIRUS.jl#50 @gdalle
Comparison with original implementation
I couldn't find the original implementation linked in any of the papers by the SIRUS authors, did I miss something?

In addition to a link, it would be nice to provide quantitative arguments for the claims found eg. in the docs:

> this Julia implementation is more easy to inspect

> this implementation is fast

https://github.com/openjournals/joss-reviews/issues/5786
