### 2023-09-07 rikhuijzer/SIRUS.jl#51 @gdalle
Performance evaluation may hint at bug?
In the paper, you state

> For the multiclass Iris classification and the Boston Housing regression datasets, the performance was worse than the other models. It could be that this is caused by a bug in the implementation or because this is a fundamental issue in the algorithm

Isn't it possible to compare with the numerical results of the original paper to validate the implementation?

https://github.com/openjournals/joss-reviews/issues/5786
