### 2023-09-07 rikhuijzer/SIRUS.jl#51 @gdalle
See also #50
