### 2023-09-07 rikhuijzer/SIRUS.jl#52 @gdalle
State of the field is missing
While the purpose of the package is clear (reimplementing a R/C++ code), I think it would benefit from a broader perspective on the state of ML and interpretability in Julia.
For instance explain what MLJ is, whether there are concurrent frameworks, and which packages offer interpretability tools (even based on black box models).

https://github.com/openjournals/joss-reviews/issues/5786
