### 2023-09-07 #5786 @gdalle
Just made a first pass, it is a really cool project with a very nice hand-holding documentation! Since I'm a bit picky, I opened a few issues on the repo, but hopefully they shouldn't be too hard to tackle. Well done @rikhuijzer!
