### 2023-09-11 5f6bbd1
Add quantitive information about lines of code
