### 2023-09-11 acee8b4
Improve binary classification tutorial
Thanks to feedback provided by Guillaume Dalle in #47.

Closes #47.

Co-authored-by: 22795598+gdalle@users.noreply.github.com
