### 2023-09-11 8153378
Set `rng` for `DecisionTree` in `test/mlj.jl`
