+++ test/mlj.jl
@@ -169,34 +169,6 @@ e = _evaluate(StableRulesClassifier(; rng=_rng(), n_trees), X, y)
 e2 = _evaluate(StableRulesClassifier(; rng=_rng(), n_trees), X, y)
 @test _score(e) == _score(e2)
 @test 0.7 < _score(e)
-
-let
-    data = "titanic"
-    hyper = (;)
-
-    e = _evaluate!(results, data, LogisticClassifier, hyper)
-
-    e = _evaluate!(results, data, XGBoostClassifier, hyper)
-
-    hyper = (; max_depth=2)
-    e = _evaluate!(results, data, XGBoostClassifier, hyper)
-
-    hyper = (;)
-    e = _evaluate!(results, data, DecisionTreeClassifier, hyper)
-
-    hyper = (; rng=_rng(), max_depth=2)
-    e = _evaluate!(results, data, StableForestClassifier, hyper)
-    @test 0.80 < _score(e)
-
-    hyper = (; rng=_rng(), max_depth=2, max_rules=30)
-    e = _evaluate!(results, data, StableRulesClassifier, hyper)
-    @test 0.79 < _score(e)
-
-    hyper = (; rng=_rng(), max_depth=2, max_rules=10)
-    e = _evaluate!(results, data, StableRulesClassifier, hyper)
-    @test 0.79 < _score(e)
-end
-
 @testset "feature names" begin
     rng = _rng()
     n = 50
@@ -209,66 +181,18 @@ end
 end
 
 let
-    data = "cancer"
-    measure = auc
-
-    hyper = (;)
-    e = _evaluate!(results, data, MultinomialClassifier, hyper; measure)
-
-    hyper = (;)
-    e = _evaluate!(results, data, XGBoostClassifier, hyper; measure)
-
-    hyper = (; max_depth=2)
-    e = _evaluate!(results, data, XGBoostClassifier, hyper; measure)
-
-    hyper = (;)
-    e = _evaluate!(results, data, DecisionTreeClassifier, hyper; measure)
-
-    hyper = (; rng=_rng(), max_depth=2)
-    e = _evaluate!(results, data, StableForestClassifier, hyper; measure)
-
-    hyper = (; rng=_rng(), max_depth=2, max_rules=30)
-    e = _evaluate!(results, data, StableRulesClassifier, hyper; measure)
-
-    hyper = (; rng=_rng(), max_depth=2, max_rules=10)
-    e = _evaluate!(results, data, StableRulesClassifier, hyper; measure)
-end
-
-let
-    data = "diabetes"
-    hyper = (;)
-
-    e = _evaluate!(results, data, LogisticClassifier, hyper)
-
-    e = _evaluate!(results, data, XGBoostClassifier, hyper)
-
-    hyper = (; max_depth=2)
-    e = _evaluate!(results, data, XGBoostClassifier, hyper)
-
-    hyper = (;)
-    e = _evaluate!(results, data, DecisionTreeClassifier, hyper)
-
-    hyper = (; rng=_rng(), max_depth=2)
-    e = _evaluate!(results, data, StableForestClassifier, hyper)
-
-    hyper = (; rng=_rng(), max_depth=2, max_rules=30)
-    e = _evaluate!(results, data, StableRulesClassifier, hyper)
-
-    hyper = (; rng=_rng(), max_depth=2, max_rules=10)
-    e = _evaluate!(results, data, StableRulesClassifier, hyper)
-end
-
-let
     data = "haberman"
+    hyper = (; rng=_rng())
+    _evaluate!(results, data, DecisionTreeClassifier)
+
     hyper = (;)
     _evaluate!(results, data, LogisticClassifier, hyper)
 
+    hyper = (;)
[…]
