### 2023-09-12 a148080
Force a stable sorting algorithm
