### 2023-09-13 2ee4204
Remove SIMD
