### 2023-09-13 367728e
Put SIMD back
