### 2023-09-13 73bec77
Ensure latest stable Julia version for Docs build
