+++ .github/workflows/Docs.yml
@@ -19,6 +19,8 @@ jobs:
         with:
           version: '1'
       - uses: julia-actions/cache@v1
+        with:
+          cache-name: 'docs'
       - uses: julia-actions/julia-buildpkg@v1
       - uses: julia-actions/julia-docdeploy@v1
         env:
