### 2023-09-13 2b8998a
Print Julia version in GitHub summary
