+++ paper/paper.bib
@@ -157,8 +157,7 @@ URL={https://doi.org/10.1214/20-EJS1792}
 @book{molnar2022interpretable,
   title={Interpretable machine learning},
   author={Molnar, Christoph},
-  year={2022},
-  publisher={Lulu.com}
+  year={2022}
 }
 
 @software{sadeghi2022decisiontree,
