### 2023-09-13 rikhuijzer/SIRUS.jl#54 @sylvaticus
Code issues in the paper
Hello, I am trying to replicate the example in the JOSS paper but I have some issues.

1. First to reproduce it I need to use the whole `register_haberman()` and ` _haberman()` functions. This data processing step distracts from the objective of showing SIRUS, so perhaps you can find some other simple way to retrieve an already-nicelly-formatted dataset for work on it, or you may want to put `X` and `y` already nicely formatted somewhere that you (and the reader) can retrieve them with a `HTTP.get` or a `Downloads.download`...

2. You are mixing MLJ and non-MLJ API together and I think this is distracting.. for example `model = StableRulesClassifier(; max_depth=2, max_rules=8)` is not MLJ API.
Instead it should be:
```julia
modelType   = @load StableRulesClassifier pkg = "SIRUS"
model       = modelType()
```
3. Related to the previous point, I can't find in the MLJ constructor the option `max_rule` that you describe in the paper
4. The `mach.fitresult` call to me doesn't return the nice rule base that you present in the paper but rather a ugly:
```julia
julia> mach.fitresult
SIRUS.StableForest{Float64}(Union{SIRUS.Leaf, SIRUS.Node}[SIRUS.ClassificationLeaf([0.5, 0.5, 0.0]), SIRUS.ClassificationLeaf([0.5, 0.0, 0.5]), SIRUS.ClassificationLeaf([0.0, 0.5, 0.5]), SIRUS.ClassificationLeaf([0.0, 0.0, 1.0]), SIRUS.ClassificationLeaf([0.0, 0.0, 1.0]), SIRUS.ClassificationLeaf([0.5, 0.0, 0.5]), SIRUS.ClassificationLeaf([0.5, 0.0, 0.5]), SIRUS.ClassificationLeaf([0.5, 0.5, 0.0]), SIRUS.ClassificationLeaf([1.0, 0.0, 0.0]), SIRUS.ClassificationLeaf([1.0, 0.0, 0.0])  …  SIRUS.ClassificationLeaf([0.5, 0.5, 0.0]), SIRUS.ClassificationLeaf([0.5, 0.5, 0.0]), SIRUS.ClassificationLeaf([0.0, 1.0, 0.0]), SIRUS.ClassificationLeaf([0.5, 0.5, 0.0]), SIRUS.ClassificationLeaf([0.5, 0.0, 0.5]), SIRUS.ClassificationLeaf([0.5, 0.0, 0.5]), SIRUS.ClassificationLeaf([0.0, 0.0, 1.0]), SIRUS.ClassificationLeaf([0.5, 0.0, 0.5]), SIRUS.ClassificationLeaf([0.5, 0.0, 0.5]), SIRUS.ClassificationLeaf([0.0, 1.0, 0.0])], SIRUS.Classification(), [0.0, 1.0, 2.0])
```
5. I have tried to apply the model to the iris dataset but I got predictions that don't even define a probability distribution (bug for multiclasses??):

```julia
julia> (X,y)       = @load_iris;
julia> modelType   = @load StableRulesClassifier pkg = "SIRUS";
[ Info: For silent loading, specify `verbosity=0`.
import SIRUS ✔
julia> model       = modelType() ;
julia> mach        = machine(model, X, y);
julia> fit!(mach);
[ Info: Training machine(StableForestClassifier(rng = Random.TaskLocalRNG(), …), …).
[ Info: Converting outcome classes ["setosa", "versicolor", "virginica"] to [0, 1, 2].
julia> est_classes = predict(mach, X)
150-element CategoricalDistributions.UnivariateFiniteVector{Multiclass{3}, Float64, UInt8, Float64}:
 UnivariateFinite{Multiclass{3}}(0.0=>0.5, 1.0=>0.5, 2.0=>0.5)
 UnivariateFinite{Multiclass{3}}(0.0=>0.5, 1.0=>0.5, 2.0=>0.5)
 UnivariateFinite{Multiclass{3}}(0.0=>0.5, 1.0=>0.5, 2.0=>0.5)
```

I believe that it is useful to show an example that for as much concise and simple as possible, it is also reproducible.

https://github.com/openjournals/joss-reviews/issues/5786
