### 2023-09-15 73e4986
Add a basic example to the docs
