+++ docs/src/basic-example.jl
@@ -1,0 +1,199 @@
+### A Pluto.jl notebook ###
+# v0.19.27
+
+using Markdown
+using InteractiveUtils
+
+# ╔═╡ 8ca5e9f4-539c-11ee-0b5a-ab77d2a5bfbf
+# ╠═╡ show_logs = false
+begin
+	ENV["DATADEPS_ALWAYS_ACCEPT"] = "true"
+
+	PKGDIR = dirname(dirname(@__DIR__))
+    DOCS_DIR = dirname(@__DIR__)
+	using Pkg: Pkg
+	Pkg.activate(DOCS_DIR)
+	Pkg.develop(; path=PKGDIR)
+end
+
+# ╔═╡ 1c1bd75a-9266-4256-bfea-ad60dd1c1d1c
+begin
+	using CategoricalArrays: categorical
+	using CSV: CSV
+	using DataDeps: DataDeps, DataDep, @datadep_str
+	using DataFrames
+	using MLJ
+	using PlutoUI: TableOfContents # hide
+	using StableRNGs: StableRNG
+	using SIRUS: StableRulesClassifier
+end
+
+# ╔═╡ 3071aa12-92df-47b2-a5ce-a54a7110ab6a
+md"""
+# Basic Example
+
+This page shows a basic example for using SIRUS.jl on a dataset via the Machine Learning Julia (MLJ.jl) interface.
+"""
+
+# ╔═╡ 0b4f89c4-ccb0-4cc9-b7bb-3f630ba2398c
+md"""
+To show the algorithm, we'll use Haberman's survival dataset.
+We load it via `DataDeps.jl` so that we can use a checksum for verification and to cache the dataset.
+"""
+
+# ╔═╡ 75550619-b310-4c66-9371-93656f78765c
+# ╠═╡ show_logs = false
+let
+    name = "Haberman"
+    message = "Haberman's Survival Data Set"
+    remote_path = "https://github.com/rikhuijzer/haberman-survival-dataset/releases/download/v1.0.0/haberman.csv"
+    checksum = "a7e9aeb249e11ac17c2b8ea4fdafd5c9392219d27cb819ffaeb8a869eb727a0f"
+    DataDeps.register(DataDep(name, message, remote_path, checksum))
+end;
+
+# ╔═╡ 05aaa007-0fe0-44ef-b815-ecf9e5f474f7
+md"""
+After dataset registration, we can load it into a `DataFrame`:
+"""
+
+# ╔═╡ ac2d7dbc-364f-437f-b66f-8eb288395275
+data = let
+    dir = datadep"Haberman"
+    path = joinpath(dir, "haberman.csv")
+    df = CSV.read(path, DataFrame)
+    df[!, :survival] = categorical(df.survival)
+    df
+end
+
+# ╔═╡ 08a4ca2b-bc65-4c29-9528-f4789272143a
+md"And split it into features (`X`) and outcomes (`y`):"
+
+# ╔═╡ e037d952-e489-41b6-afc9-317a8c17e6c4
+X = select(data, Not(:survival));
+
+# ╔═╡ 2f921f63-5148-4726-9839-c84217f60e0b
+y = data.survival;
+
+# ╔═╡ a1764625-4b7a-42f3-9e61-3d26122d86da
+md"""
+Next, we can load the model that we want to use.
+Since Haberman's outcome column (`survival`) contains 0's and 1's, we use the `StableRulesClassifier`.
+We can make the `StableRulesClassifier` symbol available via MLJ's `@load`:
+```julia
+StableRulesClassifier = @load StableRulesClassifier pkg="SIRUS"
[…]
