### 2023-09-15 b54a901
Update code example and output for the fitted model
