### 2023-09-18 rikhuijzer/SIRUS.jl#49 @rikhuijzer
This is a nice suggestion. I've added one at https://sirus.jl.huijzer.xyz/dev/basic-example/. I'll also add this to the paper. @gdalle, can you close this issue if you find the example sufficiently clear? I'm adding the example also to the paper and will request another PDF generate once I'm done with updating the paper.
