### 2023-09-18 rikhuijzer/SIRUS.jl#55 @sylvaticus
Various doc issues
Hello, some suggestion concerning the documentation.

- You could  remove the quotes in the table of page https://sirus.jl.huijzer.xyz/dev/binary-classification/
- I still think (personal and debatable  opinion) that  the "basic example" still looks more complex than it should be. If an user run `PKGDIR = dirname(dirname(@__DIR__))` or `DOCS_DIR = dirname(@__DIR__)`, it  would find itself in a different directory than the one in the documentation script. Also, users are not supposed to develop a package in order to use it.

  If,  for example, you add the (small) dataset to the repository as  `[SIRUS_FOLDER]/test/data/haberman.csv` you could simply use something like that as minimal example for a user:

  ```julia
  using CSV, DataFrames, CategoricalArrays, MLJ, SIRUS
  data_filename = joinpath(dirname(pathof(SIRUS)),"..","test","data","haberman.csv")
  data = CSV.read(data_filename, DataFrame)
  X    = select(data, Not(:survival))
  y    = categorical(data.survival)

  model = StableRulesClassifier(max_depth=2, max_rules=8)
  mach  = machine(model, X, y)
  MLJ.fit!(mach)
  mach.fitresult
  ```

  ```julia
  predictions = predict(mach, X)
  resampling = CV(nfolds=10, shuffle=true);
  evaluate(model, X, y; resampling, measure=auc)
  ```
  This would also have the advantage that you could use the dataset easily in your package tests (without download from a remote server).
  I have also removed the StableRng as I think replicability is not the focus of a minimal example and the `begin` and `let` blocks.

- I think some content of "Advanced example" could go to "Implementation overview". Also "Benchmarks" could perhaps go on its own section so that "advanced example" could focus on how to use the various options offered by the model (the hyperparameters)

https://github.com/openjournals/joss-reviews/issues/5786
