### 2023-09-18 rikhuijzer/SIRUS.jl#48 @gdalle
lgtm
