### 2023-09-18 rikhuijzer/SIRUS.jl#50 @gdalle
very cool!
