### 2023-09-18 rikhuijzer/SIRUS.jl#49 @gdalle
yes the example is good!
