+++ paper/paper.bib
@@ -1,3 +1,14 @@
+@article{ashari2013performance,
+  title={Performance comparison between Na{\"\i}ve Bayes, decision tree and k-nearest neighbor in searching alternative design in an energy simulation tool},
+  author={Ashari, Ahmad and Paryudi, Iman and Tjoa, A Min},
+  journal={International Journal of Advanced Computer Science and Applications},
+  volume={4},
+  number={11},
+  year={2013},
+  publisher={Citeseer},
+  doi={10.14569/IJACSA.2013.041105}
+}
+
 @article{barredo2020explainable,
   title={{Explainable Artificial Intelligence (XAI)}: Concepts, taxonomies, opportunities and challenges toward responsible {AI}},
   author={Barredo Arrieta, Alejandro and D{\'i}az-Rodr{\'i}guez, Natalia and Del Ser, Javier and Bennetot, Adrien and Tabik, Siham and Barbado, Alberto and Garc{\'i}a, Salvador and Gil-L{\'o}pez, Sergio and Molina, Daniel and Benjamins, Richard and others},
@@ -160,6 +171,43 @@ URL={https://doi.org/10.1214/20-EJS1792}
   year={2022}
 }
 
+@software{grisel2023scikit,
+  author={Olivier Grisel and
+          Andreas Mueller and
+          Lars and
+          Alexandre Gramfort and
+          Gilles Louppe and
+          Thomas J. Fan and
+          Peter Prettenhofer and
+          Mathieu Blondel and
+          Vlad Niculae and
+          Joel Nothman and
+          Arnaud Joly and
+          Guillaume Lemaitre and
+          Loïc Estève and
+          Jake Vanderplas and
+          Jérémie du Boisberranger and
+          manoj kumar and
+          Hanmin Qin and
+          Nicolas Hug and
+          Nelle Varoquaux and
+          Robert Layton and
+          Adrin Jalali and
+          Jan Hendrik Metzen and
+          (Venkat) Raghav, Rajagopalan and
+          Johannes Schönberger and
+          Roman Yurchak and
+          Julien Jerphanion and
+          Tom Dupré la Tour and
+          Wei Li and
+          Lucy Liu and
+          Chiara Marmo},
+  title={scikit-learn/scikit-learn: Scikit-learn},
+  publisher={Zenodo},
+  doi={10.5281/zenodo.8363803},
+  url={https://doi.org/10.5281/zenodo.8363803}
+}
+
 @software{sadeghi2022decisiontree,
   title={{DecisionTree.jl} - A {Julia} implementation of the {CART} {Decision Tree and Random Forest} algorithms},
   author={Ben Sadeghi and Poom Chiarawongse and Kevin Squire and Daniel C. Jones and Andreas Noack and Cédric St-Jean and Rik Huijzer and Roland Schätzle and Ian Butterworth and Yu-Fong Peng and Anthony Blaom},
@@ -185,6 +233,49 @@ URL={https://doi.org/10.1214/20-EJS1792}
   journal={arXiv preprint arXiv:2001.10488},
   year={2020},
   doi={10.48550/arXiv.2001.10488}
+}
+
+@article{innes2018flux,
+  title = {Flux: Elegant machine learning with Julia},
+  author={Mike Innes},
+  journal = {Journal of Open Source Software},
+  doi={10.21105/joss.00602},
+  url={https://doi.org/10.21105/joss.00602},
+  year={2018},
+  publisher = {The Open Journal},
+  volume = {3},
+  number = {25},
+  pages = {602}
+}
+
+@article{ke2017lightgbm,
+  title={{LightGBM}: A highly efficient gradient boosting decision tree},
[…]
