+++ paper/paper.bib
@@ -204,6 +204,7 @@ URL={https://doi.org/10.1214/20-EJS1792}
           Chiara Marmo},
   title={scikit-learn/scikit-learn: Scikit-learn},
   publisher={Zenodo},
+  year={2023},
   doi={10.5281/zenodo.8363803},
   url={https://doi.org/10.5281/zenodo.8363803}
 }
