### 2023-09-26 996702b
Make the introduction flow better
