### 2023-09-26 2b13aba
Set hideall for the basic example noteboko
