### 2023-09-26 21d6ae6
Get rid of quotation marks in table
