### 2023-09-26 cab8250
Print models in binary classification example better
