+++ paper/paper.md
@@ -22,7 +22,7 @@ bibliography: paper.bib
 
 # Summary
 
-SIRUS.jl[^1] is a pure Julia implementation of the original Stable and Interpretable RUle Sets (SIRUS) algorithm.
+SIRUS.jl[^1] is an implementation of the original Stable and Interpretable RUle Sets (SIRUS) algorithm in the Julia programming language.
 The SIRUS algorithm is a fully interpretable version of random forests, that is, it reduces thousands of trees in the forest to a much lower number of interpretable rules (e.g., 10 or 20).
 With our Julia implementation, we aimed to reproduce the original C++ and R implementation in a high-level language to verify the algorithm as well as making the code easier to read.
 We show that the model performs well on classification tasks while retaining interpretability and stability.
@@ -32,24 +32,28 @@ In turn, this allows others to research the algorithm further or easily port it 
 # Statement of need
 
 Many of the modern day machine learning models are noninterpretable models, also known as _black box_ models.
-Well-known examples of noninterpretable models are random forests [@breiman2001random] and neural networks, available in Julia via, for example, LightGBM.jl [@ke2017lightgbm], Flux.jl [@innes2018flux], and BetaML.jl [@lobianco2021betaml].
-Although these models are very capable and commonly used, they can be problematic in high stakes domains where model decisions have real-world impact on individuals, such as suggesting treatments or selecting personnel.
-Such noninterpretable models may lead to unsafe, unfair, or unreliable predictions [@doshi2017towards; @barredo2020explainable].
+Well-known examples of noninterpretable models are random forests [@breiman2001random] and neural networks.
+Such models are available in the Julia programming language via, for example, LightGBM.jl [@ke2017lightgbm], Flux.jl [@innes2018flux], and BetaML.jl [@lobianco2021betaml].
+Although these models can obtain high predictive performance and are commonly used, they can be problematic in high stakes domains where model decisions have real-world impact on individuals, such as suggesting treatments or selecting personnel.
+The reason is that noninterpretable models may lead to unsafe, unfair, or unreliable predictions [@doshi2017towards; @barredo2020explainable].
+Furthermore, interpretable models may allow researchers to learn more from the model, which in turn may allow researchers to make better model decisions and achieve a higher predictive performance.
+
 However, the set of interpretable models is often limited to ordinary and generalized regression models, decision trees, RuleFit, naive Bayes classification, and k-nearest neighbors [@molnar2022interpretable].
-For these models, predictive performance can be poor.
+For these models, however, predictive performance can be poor for certain tasks.
 Linear models, for instance, may perform poor when features are correlated and can be sensitive to the choice of hyperparameters.
 For decision trees, predictive performance is poor compared to random forests [@james2013introduction].
-RuleFit is not available in Julia and is _unstable_, which will be discussed later.
-Naive Bayes, available in Julia as NaiveBayes.jl[^2], is often overlooked and can be very suitable for many situations, but does require features to be independent [@ashari2013performance].
-Random forests outperform linear models and decision trees and are often considered interpretable.
-However, they are not fully interpretable due to the large number of trees, typically thousands, in the forests.
-Model interpretation techniques, such as SHAP [@lundberg2017unified] or Shapley, available via Shapley.jl[^5], do not solve this as they do not clearly explain predictions made by the models.
-This is because model interpretation techniques convert the complex model to a simplified representation.
+RuleFit is not available in Julia and is _unstable_ [@benard2021sirus], meaning sensitive to small changes in data.
+Naive Bayes, available in Julia as NaiveBayes.jl[^2], is often overlooked and can be a suitable solution, but only if the features are independent [@ashari2013performance].
+
[…]
