### 2023-09-27 #5786 @rikhuijzer
Thanks all for the patience. Above is a new version of the PDF. In summary, this updated PDF should solve many of the raised issues. We have described more commonly-used packages and how they related to SIRUS, as was raised in https://github.com/rikhuijzer/SIRUS.jl/issues/52. Furthermore, we have improved the example in the paper as well as in the documentation (<https://sirus.jl.huijzer.xyz/dev/basic-example/>), as was raised in https://github.com/rikhuijzer/SIRUS.jl/issues/54 and https://github.com/rikhuijzer/SIRUS.jl/issues/55. I'll keep working on the remaining issues, in particular https://github.com/rikhuijzer/SIRUS.jl/issues/51.
