### 2023-09-27 rikhuijzer/SIRUS.jl#55 @rikhuijzer
> You could remove the quotes in the table of page https://sirus.jl.huijzer.xyz/dev/binary-classification/

Done.

> I still think (personal and debatable opinion) that the "basic example" still looks more complex than it should be. If an user run PKGDIR = dirname(dirname(@__DIR__)) or DOCS_DIR = dirname(@__DIR__), it would find itself in a different directory than the one in the documentation script. Also, users are not supposed to develop a package in order to use it.

Fixed.

> If, for example, you add the (small) dataset to the repository as [SIRUS_FOLDER]/test/data/haberman.csv you could simply use something like that as minimal example for a user:

I have considered this, but am hesitant to add data to the source code. I have also considered a less heavy approach via the Julia (Lazy) Artifact system. However, the main issue I have is with the question: If we add a dataset, shouldn't we add more datasets? I think no dataset is a requirement for the package and therefore they should not be linked.

> This would also have the advantage that you could use the dataset easily in your package tests (without download from a remote server).

Datadeps does cache the file, so in many cases the test will not require a download. DataDeps is not yet cached via the GitHub Actions Cache (https://github.com/julia-actions/cache/issues/40), but that might be fixed in the future.

> I have also removed the StableRng as I think replicability is not the focus of a minimal example and the begin and let blocks.

Thanks for providing the example code, which made the example easy to understand. I have removed the `begin` and `let` blocks and replaced them by functions. Functions are often considered as superior to comments in code as they provide clarity while ensuring that the desciption matches the code.

> I think some content of "Advanced example" could go to "Implementation overview". Also "Benchmarks" could perhaps go on its own section so that "advanced example" could focus on how to use the various options offered by the model (the hyperparameters)

I haven't done this yet, but start working on it.
