### 2023-09-27 rikhuijzer/SIRUS.jl#52 @rikhuijzer
This was indeed a big issue. We had failed to mention many packages. Thank you for reporting it and providing some suggestions. In the latest PDF version in the REVIEW issue, we have tried to find a balance between your suggestions here and the original checklist item in the JOSS review checklist. Can you check whether the current PDF version has solved this issue?
