### 2023-09-30 913ace9
CompatHelper: bump compat for MLJBase to 1, (keep existing compat)
