### 2023-10-01 a7328dc
Force MLJBase to 1 in test compat
