+++ test/Project.toml
@@ -23,7 +23,7 @@ DataDeps = "0.7"
 DataFrames = "1.3"
 DecisionTree = "0.12"
 MLDatasets = "0.7"
-MLJBase = "0.21"
+MLJBase = "1"
 MLJDecisionTreeInterface = "0.4"
 MLJLinearModels = "0.9"
 MLJTestInterface = "0.2"
