+++ docs/Project.toml
@@ -28,7 +28,7 @@ Documenter = "0.27"
 EvoTrees = "0.16"
 LiveServer = "1"
 MLJ = "0.19"
-MLJBase = "0.21"
+MLJBase = "0.21, 1"
 MLJDecisionTreeInterface = "0.4"
 PlutoStaticHTML = "6"
 PlutoUI = "0.7"
