### 2023-10-02 c0b1028
CompatHelper: bump compat for MLJBase to 1, (keep existing compat) (#56)
This pull request changes the compat entry for the `MLJBase` package
from `0.21` to `0.21, 1`.
This keeps the compat entries for earlier versions.

---------

Co-authored-by: Rik Huijzer <github@huijzer.xyz>
