+++ Project.toml
@@ -1,7 +1,7 @@
 name = "SIRUS"
 uuid = "cdeec39e-fb35-4959-aadb-a1dd5dede958"
 authors = ["Rik Huijzer <github@huijzer.xyz>"]
-version = "1.3.2"
+version = "1.3.3"
 
 [deps]
 AbstractTrees = "1520ce14-60c1-5f80-bbc7-55ef81b5835c"
@@ -20,7 +20,7 @@ Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 AbstractTrees = "0.3, 0.4"
 CategoricalArrays = "0.10"
 InlineStrings = "1"
-MLJBase = "0.21"
+MLJBase = "0.21, 1"
 MLJLinearModels = "0.8, 0.9"
 MLJModelInterface = "1.4"
 OrderedCollections = "1.6.2"
