+++ test/Project.toml
@@ -13,6 +13,7 @@ MLJTestInterface = "72560011-54dd-4dc2-94f3-c5de45b75ecd"
 MLJXGBoostInterface = "54119dfa-1dab-4055-a167-80440f4f7a91"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 StableRNGs = "860ef19b-820b-49d6-a774-d7a799459cd3"
+StatisticalMeasures = "a19d573c-0a75-4610-95b3-7071388c7541"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
@@ -23,10 +24,11 @@ DataDeps = "0.7"
 DataFrames = "1.3"
 DecisionTree = "0.12"
 MLDatasets = "0.7"
-MLJBase = "0.21"
+MLJBase = "1"
 MLJDecisionTreeInterface = "0.4"
 MLJLinearModels = "0.9"
 MLJTestInterface = "0.2"
 MLJXGBoostInterface = "0.3.8"
 StableRNGs = "1"
+StatisticalMeasures = "0.1"
 Tables = "1.7"
