+++ docs/Project.toml
@@ -24,11 +24,11 @@ CategoricalArrays = "0.10"
 DataDeps = "0.7"
 DataFrames = "1"
 DecisionTree = "0.12"
-Documenter = "0.27"
+Documenter = "1"
 EvoTrees = "0.16"
 LiveServer = "1"
-MLJ = "0.19"
-MLJBase = "0.21, 1"
+MLJ = "0.20"
+MLJBase = "1"
 MLJDecisionTreeInterface = "0.4"
 PlutoStaticHTML = "6"
 PlutoUI = "0.7"
