### 2023-10-02 a3ba815
Use `warnonly` instead of `strict`
