### 2023-10-02 a353bcc
Update docs dependencies (#57)
