+++ docs/src/basic-example.jl
@@ -1,5 +1,5 @@
 ### A Pluto.jl notebook ###
-# v0.19.27
+# v0.19.28
 
 using Markdown
 using InteractiveUtils
