+++ test/rcall.jl
@@ -1,0 +1,9 @@
+using RCall
+
+@rlibrary sirus
+
+r = R"rnorm(10)"
+
+@show r
+
+R"sessionInfo()"
