+++ test/Project.toml
@@ -11,6 +11,7 @@ MLJDecisionTreeInterface = "c6f25543-311c-4c74-83dc-3ea6d1015661"
 MLJLinearModels = "6ee0df7b-362f-4a72-a706-9e79364fb692"
 MLJTestInterface = "72560011-54dd-4dc2-94f3-c5de45b75ecd"
 MLJXGBoostInterface = "54119dfa-1dab-4055-a167-80440f4f7a91"
+RCall = "6f49c342-dc21-5d91-9882-a32aef131414"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 StableRNGs = "860ef19b-820b-49d6-a774-d7a799459cd3"
 StatisticalMeasures = "a19d573c-0a75-4610-95b3-7071388c7541"
@@ -29,6 +30,7 @@ MLJDecisionTreeInterface = "0.4"
 MLJLinearModels = "0.9"
 MLJTestInterface = "0.2"
 MLJXGBoostInterface = "0.3.8"
+RCall = "0.13"
 StableRNGs = "1"
 StatisticalMeasures = "0.1"
 Tables = "1.7"
