### 2023-10-04 b815e83
Set VERSION minimum
