+++ .github/workflows/Docs.yml
@@ -1,47 +1,0 @@
-name: Docs
-
-on:
-  push:
-    branches:
-      - main
-  pull_request:
-  workflow_dispatch:
-
-jobs:
-  BuildDocs:
-    permissions:
-      contents: write
-    runs-on: ubuntu-latest
-    timeout-minutes: 40
-    steps:
-      - uses: actions/checkout@v3
-      - uses: julia-actions/setup-julia@v1
-        with:
-          version: '1'
-      - uses: julia-actions/cache@v1
-        with:
-          cache-name: 'docs'
-      - uses: julia-actions/julia-buildpkg@v1
-      - uses: julia-actions/julia-docdeploy@v1
-        env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-
-  ConfigureDocs:
-    needs: BuildDocs
-    if: github.ref == 'refs/heads/main'
-    permissions:
-      contents: write
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v3
-        with:
-          ref: 'docs-output'
-      - run: echo "sirus.jl.huijzer.xyz" > CNAME
-      - run: echo "<meta http-equiv=\"refresh\" content=\"0; url=./dev/\"/>" > index.html
-      - uses: peaceiris/actions-gh-pages@v3
-        with:
-          force_orphan: true
-          github_token: ${{ secrets.GITHUB_TOKEN }}
-          publish_dir: .
-          publish_branch: 'docs-output'
-
