### 2023-10-04 91409d8
Can train sirus locally
