+++ test/rcall.jl
@@ -1,9 +1,62 @@
 using RCall
 
-@rlibrary sirus
+import MLJModelInterface:
+    MLJModelInterface,
+    fit,
+    predict,
+    metadata_model,
+    metadata_pkg
 
-r = R"rnorm(10)"
+using MLJModelInterface:
+    MLJModelInterface,
+    UnivariateFinite,
+    Continuous,
+    Count,
+    Deterministic,
+    Finite,
+    Probabilistic,
+    Table
 
-@show r
+const MMI = MLJModelInterface
 
-R"sessionInfo()"
+# @rlibrary sirus
+
+MMI.@mlj_model mutable struct RSirusRegressor <: Deterministic
+end
+
+R"library('sirus')"
+
+n = 100
+A = rand(_rng(), n)
+B = rand(_rng(), n)
+X = DataFrame(; A, B)
+y = rand(_rng(), n)
+
+function fit(
+        model::RSirusRegressor,
+        verbosity::Int,
+        X,
+        y
+    )
+    
+    sfit = R"""
+        sirus.fit(
+            $X,
+            $y,
+            type="auto",
+            num.rule=10,
+            p0=NULL,
+            num.rule.max = 10,
+            q=4,
+            max.depth=2,
+            num.trees=NULL,
+            num.threads=1,
+            verbose=TRUE,
+            seed=1
+        )
+    """
+end
+
+model = RSirusRegressor()
+mach = machine(model, X, y)
+fit!(mach)
