### 2023-10-04 28a4070
Implement regression wrapper
