+++ test/mlj.jl
@@ -83,12 +83,6 @@ function _with_trailing_zero(score::Real)::String
     end
 end
 
-function _evaluate(model, X, y, nfolds=10, measure=auc)
-    resampling = CV(; nfolds, shuffle=true, rng=_rng())
-    acceleration = MLJBase.CPUThreads()
-    evaluate(model, X, y; acceleration, verbosity=0, resampling, measure)
-end
-
 results = DataFrame(;
         Dataset=String[],
         Model=String[],
