### 2023-10-06 5c87eda
Test sirus via RCall (#58)
