### 2023-10-06 a271976
Add comparison to the original SIRUS implementation
Made possibly by https://github.com/rikhuijzer/SIRUS.jl/pull/58.
