+++ paper/paper.md
@@ -110,20 +110,21 @@ One model which suffers from a low stability is a decision tree, available via D
 Similarly, linear models can be highly sensitive to correlated data and, in the case of regularized linear models, the choice of hyperparameters.
 The aforementioned RuleFit algorithm also suffers from stability issues due to the unstable combination of tree fitting and rule extraction [@benard2021sirus].
 The SIRUS algorithm solves this problem by stabilizing the trees inside the forest, and the original authors have proven the correctness of this stabilization mathematically [@benard2021sirus].
-In the rest of this paper, we will compare decision trees [@sadeghi2022decisiontree], linear models, XGBoost [@chen2016xgboost], and SIRUS on their predictive performance.
+In the rest of this paper, we will compare the predictive performance of SIRUS to the performance of decision trees [@sadeghi2022decisiontree], linear models, XGBoost [@chen2016xgboost], and the original (C++/R) SIRUS implementation [@benard2021sirus].
 The interpretability and stability are summarized in Table \ref{tab:is}.
 
 \begin{table}[h!]
 \small
 \centering
-\begin{tabular}{|l|c|c|c|c|c|}
+\begin{tabular}{|l|c|c|c|c|c|c|}
 \hline
- & \textbf{Decision Tree} & \textbf{Linear Model} & \textbf{XGBoost} & \textbf{XGBoost} & \textbf{SIRUS} \\
-& & & \textbf{\scriptsize{max depth: $\mathbb{\infty}$}} & \textbf{\scriptsize{max depth: 2}} & \textbf{\scriptsize{max depth: 2}} \\
-& & & & & \textbf{\scriptsize{max rules: 10}} \\
+ & \textbf{Decision} & \textbf{Linear} & \textbf{XGBoost} & \textbf{XGBoost} & \textbf{Original} & \textbf{SIRUS.jl} \\
+ & \textbf{Tree} & \textbf{Model} & & & \textbf{SIRUS} & \\
+& & & \textbf{\scriptsize{max depth: $\mathbb{\infty}$}} & \textbf{\scriptsize{max depth: 2}} & \textbf{\scriptsize{max depth: 2}} & \textbf{\scriptsize{max depth: 2}} \\
+& & & & & \textbf{\scriptsize{max rules: 10}} & \textbf{\scriptsize{max rules: 10}} \\
 \hline
-\textbf{Interpretability} & High & High & Medium & Medium & High \\
-\textbf{Stability} & Low & Medium & High & High & High \\
+\textbf{Interpretability} & High & High & Medium & Medium & High & High \\
+\textbf{Stability} & Low & Medium & High & High & High & High \\
 \hline
 \end{tabular}
 \caption{Summary of interpretability and stability for various models.}
@@ -137,32 +138,34 @@ To make the random forests interpretable, the large number of trees are converte
 The conversion works by converting each tree to a set of rules and then pruning the rules by removing simple duplicates and linearly dependent duplicates, see the SIRUS.jl documentation or the original paper [@benard2021interpretable] for details.
 In practice, this trade-off between between model complexity and interpretability comes at a small performance cost.
 
-To show the performance, we compared SIRUS to a decision tree, linear model, and XGBoost; similar to Table \ref{tab:is}.
[…]
