### 2023-10-06 #5786 @rikhuijzer
Issue https://github.com/rikhuijzer/SIRUS.jl/issues/51 has now been addressed as well. Let me know if there is anything else I can address.
