+++ Project.toml
@@ -21,7 +21,7 @@ AbstractTrees = "0.3, 0.4"
 CategoricalArrays = "0.10"
 InlineStrings = "1"
 MLJBase = "0.21, 1"
-MLJLinearModels = "0.8, 0.9"
+MLJLinearModels = "0.8, 0.9, 0.10"
 MLJModelInterface = "1.4"
 OrderedCollections = "1.6.2"
 Tables = "1.7"
