### 2023-10-09 rikhuijzer/SIRUS.jl#60 @gdalle
Last details in JOSS paper
L31: "perform poor" => perform poorly
L80: "will most likely survive" => is more likely to survive (the proba from the first rule is not > 0.5)
L136: repetition of "models"

Table 1: why 2 columns for XGBoost and SIRUS?

https://github.com/openjournals/joss-reviews/issues/5786
