### 2023-10-09 rikhuijzer/SIRUS.jl#55 @sylvaticus
One minor issue with the example in the documentation.
When you try to run the code as described in the example you will have the error on the line:

```StableRulesClassifier = @load StableRulesClassifier pkg="SIRUS"```
```ERROR: cannot assign a value to imported variable SIRUS.StableRulesClassifier from module Main```

Because you have already imported `StableRulesClassifier` in the first chunk of the script. Please remove `using SIRUS: StableRulesClassifier` from that first chunk.

Also, I had some errors in typing the `using..` statements in that order, I suggest to write the `using DataFrames` BEFORE the line `using DataDeps: DataDeps, DataDep, @datadep_str`
