### 2023-10-09 #5786 @sylvaticus
Fine for me..
just three very minor points:
- in the paper you still refer to class 1.0 instead of 1 ("then the patient will most likely survive (class == 1.0)"
- thank you that you cite my own machine learning toolkit (BetaML), but it is not necessary.. you can keep DecisionTree.jl as example of Decision Tree/ Random Forest package, they are way more optimized (although they don't accept missing nor unordered features)
- I would rewrite the sentence "The performance scores were taken from the GitHub Actions Job Summary124 for commit 5c87eda." to "The performance scores were taken from the [SIRUS.jl test job](https://github.com/rikhuijzer/SIRUS.jl/actions/runs/6429413860/attempts/1#summary-17458424403) that was run following commit 5c87eda using GitHub Actions."
