### 2023-10-10 9017cf5
Fix "last details" issue raised by Guillaume
Fixes https://github.com/rikhuijzer/SIRUS.jl/issues/60.
