### 2023-10-10 b1a8cfa
Use `RulesClassifier` to fix `@load`
