### 2023-10-10 ca1c02b
Get rid of the `@load` in the example
