### 2023-10-10 rikhuijzer/SIRUS.jl#55 @rikhuijzer
> I think some content of "Advanced example" could go to "Implementation overview". Also "Benchmarks" could perhaps go on its own section so that "advanced example" could focus on how to use the various options offered by the model (the hyperparameters)

Could we consider this out-of-scope for now? Then I can focus on the issues that directly affect the paper's content. I've opened an issue for this at https://github.com/rikhuijzer/SIRUS.jl/issues/62.

> `StableRulesClassifier = @load StableRulesClassifier pkg="SIRUS"`

> Because you have already imported `StableRulesClassifier` in the first chunk of the script. Please remove `using SIRUS: StableRulesClassifier` from that first chunk.

It looks like https://github.com/JuliaAI/MLJModels.jl/pull/526 hasn't fixed the issue that we ran into earlier because

```
pkg> activate --temp

pkg> add MLJModels
   Resolving package versions...
    Updating `/private/var/folders/nf/xynnllys6nl13xhj67sqgt7c0000gn/T/jl_AG8cXc/Project.toml`
  [d491faf4] + MLJModels v0.16.12
  [...]

pkg> add SIRUS
   Resolving package versions...
    Updating `/private/var/folders/nf/xynnllys6nl13xhj67sqgt7c0000gn/T/jl_AG8cXc/Project.toml`
  [cdeec39e] + SIRUS v1.3.3
  [...]

julia> using MLJModels

julia> RulesClassifier = @load StableRulesClassifier pkg="SIRUS" verbosity=0
SIRUS.MLJImplementation.StableForestClassifier
```
So I'm using the latest versions of MLJModels and SIRUS but still `StableRulesClassifier` loads the `StableForestClassifier`. This is a bug that needs further investigation (I've opened an issue at https://github.com/rikhuijzer/SIRUS.jl/issues/61). For now, I've fixed the documentation by removing the `@load` in https://github.com/rikhuijzer/SIRUS.jl/commit/ca1c02bc405a5ea2562d5e944f94d8e942ead59b.

> Also, I had some errors in typing the `using..` statements in that order, I suggest to write the `using DataFrames` BEFORE the line `using DataDeps: DataDeps, DataDep, @datadep_str`

Did you get warnings during the loading of `using DataFrames`? Could it be that DataFrames was outdated in the global Julia environment? I just ran the following without warnings after updating my global environment:

```
pkg> activate --temp

julia> begin
           using CategoricalArrays: categorical
           using CSV: CSV
           using DataDeps: DataDeps, DataDep, @datadep_str
           using DataFrames
           using MLJ
           using StableRNGs: StableRNG
           using SIRUS: StableRulesClassifier
       end
 │ Packages [CategoricalArrays, DataDeps, MLJ, StableRNGs, SIRUS] not found, but packages named [CategoricalArrays, DataDeps, MLJ, StableRNGs, SIRUS] are available from a registry.
 │ Install packages?
 │   (jl_cHDhkF) pkg> add CategoricalArrays DataDeps MLJ StableRNGs SIRUS
 └ (y/n/o) [y]:
 [...]
        Info Packages marked with ⌅ have new versions available but compatibility constraints restrict them from upgrading. To see why use `status --outdated -m`

julia>
```

Sorry, by the way, for not understanding earlier that you expect to be able to run the example from the command line. I personally almost never do data science from the command line, so that's why I didn't see it before.

I just ran the whole basic example on the command line and it should now work. Please let me know if you run into issues.
