### 2023-10-10 rikhuijzer/SIRUS.jl#60 @rikhuijzer
Thank you, Guillaume. Very well spotted. My supervisor and I both missed these.

> Table 1: why 2 columns for XGBoost and SIRUS?
