### 2023-10-10 16eae13
Fix typos in Table 1
