### 2023-10-10 #5786 @rikhuijzer
> thank you that you cite my own machine learning toolkit (BetaML), but it is not necessary..

I appreciate your humility, @sylvaticus, but I think your project is a valuable mention. It was added it in response to https://github.com/rikhuijzer/SIRUS.jl/issues/52. Your paper is definitely part of the ML field in Julia.
