### 2023-10-10 #5786 @sylvaticus
@jbytecode :  Sorry, I did it yesterday but some checkboxes didn't activated  (perhaps I did click too fast). Now, there should all be checked/set... (I verify the software example yesterday)
