+++ paper/paper.bib
@@ -1,3 +1,17 @@
+@article{julia,
+  doi       = {10.1137/141000671},
+  url       = {https://doi.org/10.1137%2F141000671},
+  year      = 2017,
+  month     = {jan},
+  publisher = {Society for Industrial {\&} Applied Mathematics ({SIAM})},
+  volume    = {59},
+  number    = {1},
+  pages     = {65--98},
+  author    = {Jeff Bezanson and Alan Edelman and Stefan Karpinski and Viral B. Shah},
+  title     = {Julia: A Fresh Approach to Numerical Computing},
+  journal   = {{SIAM} Review}
+}
+
 @article{ashari2013performance,
   title={Performance comparison between Na{\"\i}ve Bayes, decision tree and k-nearest neighbor in searching alternative design in an energy simulation tool},
   author={Ashari, Ahmad and Paryudi, Iman and Tjoa, A Min},
