### 2023-10-10 #5786 @jbytecode
@rikhuijzer - The paper and bibtex seem quite good. I have added a citation for Julia in the PR (https://github.com/rikhuijzer/SIRUS.jl/pull/63). Please review and merge. Ping me when you are done. Thank you in advance.
