### 2023-10-11 07ef4e3
add citation of Julia (#63)
Due to the JOSS submission: 

https://github.com/openjournals/joss-reviews/issues/5786
