### 2023-10-11 #5786 @rikhuijzer
Thank you for the instructions, @jbytecode. I have followed them together with <https://joss.readthedocs.io/en/latest/editing.html#after-reviewers-recommend-acceptance>.

This is the Zenodo link for the record with the right paper title and author list: <https://zenodo.org/record/8398350>.

DOI: 10.5281/zenodo.8398350 (<https://doi.org/10.5281/zenodo.8398350>)
Version: 1.3.3

I hope I did everything correctly. Let me know if not. I'll also trigger a new PDF build to get the Julia citation (https://github.com/rikhuijzer/SIRUS.jl/pull/63) into the article proof:
