### 2023-10-11 7740987
update bibtex
