+++ paper/paper.bib
@@ -13,7 +13,7 @@
 }
 
 @article{ashari2013performance,
-  title={Performance comparison between Na{\"\i}ve Bayes, decision tree and k-nearest neighbor in searching alternative design in an energy simulation tool},
+  title={Performance comparison between Na{\"\i}ve {B}ayes, decision tree and k-nearest neighbor in searching alternative design in an energy simulation tool},
   author={Ashari, Ahmad and Paryudi, Iman and Tjoa, A Min},
   journal={International Journal of Advanced Computer Science and Applications},
   volume={4},
@@ -68,7 +68,7 @@ URL={https://doi.org/10.1214/20-EJS1792}
 }
 
 @article{blaom2020mlj,
-  title={{MLJ}: A Julia package for composable machine learning},
+  title={{MLJ}: A {J}ulia package for composable machine learning},
   author={Anthony D. Blaom and Franz Kiraly and Thibaut Lienart and Yiannis Simillides and Diego Arenas and Sebastian J. Vollmer},
   year={2020},
   publisher={The Open Journal},
@@ -251,7 +251,7 @@ URL={https://doi.org/10.1214/20-EJS1792}
 }
 
 @article{innes2018flux,
-  title = {Flux: Elegant machine learning with Julia},
+  title = {Flux: Elegant machine learning with {J}ulia},
   author={Mike Innes},
   journal = {Journal of Open Source Software},
   doi={10.21105/joss.00602},
@@ -272,7 +272,7 @@ URL={https://doi.org/10.1214/20-EJS1792}
 }
 
 @article{lobianco2021betaml,
-  title={BetaML: The Beta Machine Learning Toolkit, a self-contained repository of Machine Learning algorithms in Julia},
+  title={BetaML: The Beta Machine Learning Toolkit, a self-contained repository of Machine Learning algorithms in {J}ulia},
   author={Antonello Lobianco},
   doi={10.21105/joss.02849},
   url={https://doi.org/10.21105/joss.02849},
