### 2023-10-11 #5786 @jbytecode
@rikhuijzer - I have just sent a pull request (https://github.com/rikhuijzer/SIRUS.jl/pull/64) for some minor changes / corrections in the bibliography file. Could you please review and merge? Thank you in advance.
