### 2023-10-11 3767dc5
Update bibtex (#64)
Due to the JOSS submission:
https://github.com/openjournals/joss-reviews/issues/5786
