### 2023-10-11 #5786 @jbytecode
@rikhuijzer - The paper, the bibliography, and the software repository look good to me. I am now recommending an acceptance. Our track editor will make the final decision. Thank you for submitting your work to JOSS.

@sylvaticus, @gdalle - Thank you for reviewing and your valuable thoughts. Hope we work in future submissions again.
