### 2023-10-12 #5786 @arfon
> Actually @jbytecode I was thinking of answering the JOSS call for editors

@gdalle – please go ahead and apply, no need to email me directly about this!
